remove the v1 drawer and rename v2 drawer (#1982)
* remove the v1 drawer and rename v2 drawer

* remove commented out code
