+++ frontends/ol-components/src/components/LearningResourceExpanded/InfoSection.test.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { render, screen, within } from "@testing-library/react"
 import { courses } from "../LearningResourceCard/testUtils"
-import InfoSectionV2 from "./InfoSectionV2"
+import InfoSectionV2 from "./InfoSection"
 import { ThemeProvider } from "../ThemeProvider/ThemeProvider"
 import { formatRunDate } from "ol-utilities"
 import invariant from "tiny-invariant"
