+++ frontends/ol-components/src/components/LearningResourceExpanded/InfoSection.tsx
