--- frontends/ol-components/src/components/LearningResourceExpanded/InfoSectionV1.test.tsx
@@ -1,99 +1,0 @@
-import React from "react"
-import { render, screen } from "@testing-library/react"
-import { courses } from "../LearningResourceCard/testUtils"
-import InfoSectionV1 from "./InfoSectionV1"
-import { ThemeProvider } from "../ThemeProvider/ThemeProvider"
-
-describe("Learning resource info section pricing", () => {
-  test("Free course, no certificate", () => {
-    render(<InfoSectionV1 resource={courses.free.noCertificate} />, {
-      wrapper: ThemeProvider,
-    })
-
-    screen.getByText("Free")
-    expect(screen.queryByText("Paid")).toBeNull()
-    expect(screen.queryByText("Earn a certificate:")).toBeNull()
-    expect(screen.queryByText("Certificate included")).toBeNull()
-  })
-
-  test("Free course, with certificate, one price", () => {
-    render(<InfoSectionV1 resource={courses.free.withCertificateOnePrice} />, {
-      wrapper: ThemeProvider,
-    })
-
-    screen.getByText("Free")
-    expect(screen.queryByText("Paid")).toBeNull()
-    screen.getByText("Earn a certificate:")
-    screen.getByText("$49")
-  })
-
-  test("Free course, with certificate, price range", () => {
-    render(
-      <InfoSectionV1 resource={courses.free.withCertificatePriceRange} />,
-      {
-        wrapper: ThemeProvider,
-      },
-    )
-
-    screen.getByText("Free")
-    expect(screen.queryByText("Paid")).toBeNull()
-    screen.getByText("Earn a certificate:")
-    screen.getByText("$49 – $99")
-  })
-
-  test("Unknown price, no certificate", () => {
-    render(<InfoSectionV1 resource={courses.unknownPrice.noCertificate} />, {
-      wrapper: ThemeProvider,
-    })
-
-    screen.getByText("Paid")
-    expect(screen.queryByText("Free")).toBeNull()
-    expect(screen.queryByText("Earn a certificate:")).toBeNull()
-    expect(screen.queryByText("Certificate included")).toBeNull()
-  })
-
-  test("Unknown price, with certificate", () => {
-    render(<InfoSectionV1 resource={courses.unknownPrice.withCertificate} />, {
-      wrapper: ThemeProvider,
-    })
-
-    screen.getByText("Paid")
-    expect(screen.queryByText("Free")).toBeNull()
-    screen.getByText("Certificate included")
-  })
-
-  test("Paid course, no certificate", () => {
-    render(<InfoSectionV1 resource={courses.paid.withoutCertificate} />, {
-      wrapper: ThemeProvider,
-    })
-
-    screen.getByText("$49")
-    expect(screen.queryByText("Paid")).toBeNull()
-    expect(screen.queryByText("Free")).toBeNull()
-    expect(screen.queryByText("Earn a certificate:")).toBeNull()
-    expect(screen.queryByText("Certificate included")).toBeNull()
-  })
-
-  test("Paid course, with certificate, one price", () => {
-    render(<InfoSectionV1 resource={courses.paid.withCerticateOnePrice} />, {
-      wrapper: ThemeProvider,
-    })
-
-    screen.getByText("$49")
-    expect(screen.queryByText("Paid")).toBeNull()
-    screen.getByText("Certificate included")
-  })
-
-  test("Paid course, with certificate, price range", () => {
-    render(
-      <InfoSectionV1 resource={courses.paid.withCertificatePriceRange} />,
-      {
-        wrapper: ThemeProvider,
-      },
-    )
-
-    screen.getByText("$49 – $99")
-    expect(screen.queryByText("Paid")).toBeNull()
-    screen.getByText("Certificate included")
-  })
-})
