+++ frontends/ol-components/src/index.ts
@@ -131,8 +131,7 @@ export * from "./components/LearningResourceCard/LearningResourceCard"
 export { LearningResourceListCard } from "./components/LearningResourceCard/LearningResourceListCard"
 export type { LearningResourceListCardProps } from "./components/LearningResourceCard/LearningResourceListCard"
 export * from "./components/LearningResourceCard/LearningResourceListCardCondensed"
-export * from "./components/LearningResourceExpanded/LearningResourceExpandedV1"
-export * from "./components/LearningResourceExpanded/LearningResourceExpandedV2"
+export * from "./components/LearningResourceExpanded/LearningResourceExpanded"
 export * from "./components/LoadingSpinner/LoadingSpinner"
 export * from "./components/Logo/Logo"
 export * from "./components/NavDrawer/NavDrawer"
